fix tampilan approved user, cafe, dan fix tampilan user beserta fungsinya

+++ database/migrations/2026_05_01_131352_create_hasils_table.php
@@ -6,29 +6,25 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
-{
-    Schema::create('tbl_hasil', function (Blueprint $table) {
-        $table->increments('id_hasil');
-        $table->string('id_alternatif', 5);
-        $table->decimal('nilai_v', 10, 6);
-        $table->integer('ranking');
+    {
+        if (!Schema::hasTable('tbl_hasil')) {
+            Schema::create('tbl_hasil', function (Blueprint $table) {
+                $table->increments('id_hasil');
+                $table->string('id_alternatif', 5);
+                $table->decimal('nilai_v', 10, 6);
+                $table->integer('ranking');
 
-        $table->foreign('id_alternatif')
-            ->references('id_alternatif')
-            ->on('tbl_alternatif')
-            ->onDelete('cascade');
-    });
-}
+                $table->foreign('id_alternatif')
+                    ->references('id_alternatif')
+                    ->on('tbl_alternatif')
+                    ->onDelete('cascade');
+            });
+        }
+    }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
-        Schema::dropIfExists('hasils');
+        Schema::dropIfExists('tbl_hasil');
     }
 };
